+++ keep-ui/app/(keep)/alerts/alert-table-server-side.tsx
@@ -1,12 +1,7 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { Table, Card } from "@tremor/react";
 import { AlertsTableBody } from "./alerts-table-body";
-import {
-  AlertDto,
-  reverseSeverityMapping,
-  Severity,
-  Status,
-} from "@/entities/alerts/model";
+import { AlertDto, reverseSeverityMapping } from "@/entities/alerts/model";
 import {
   getCoreRowModel,
   useReactTable,
@@ -170,7 +165,7 @@ export function AlertTableServerSide({
     pageSize: 20,
   });
 
-  const [viewedAlerts, setViewedAlerts] = useLocalStorage<ViewedAlert[]>(
+  const [, setViewedAlerts] = useLocalStorage<ViewedAlert[]>(
     `viewed-alerts-${presetName}`,
     []
   );
@@ -232,13 +227,13 @@ export function AlertTableServerSide({
   }, [alertsQuery, onReload]);
 
   const [tabs, setTabs] = useState([
-    { name: "All", filter: (alert: AlertDto) => true },
+    { name: "All", filter: () => true },
     ...presetTabs.map((tab) => ({
       name: tab.name,
       filter: (alert: AlertDto) => evalWithContext(alert, tab.filter),
       id: tab.id,
     })),
-    { name: "+", filter: (alert: AlertDto) => true }, // a special tab to add new tabs
+    { name: "+", filter: () => true }, // a special tab to add new tabs
   ]);
 
   const [selectedTab, setSelectedTab] = useState(0);
@@ -307,7 +302,6 @@ export function AlertTableServerSide({
 
   let showSkeleton = isAsyncLoading;
   const isTableEmpty = table.getPageCount() === 0;
-  let showEmptyState = !alertsQuery.cel && isTableEmpty && !isAsyncLoading;
   const showFilterEmptyState = isTableEmpty && !!filterCel;
   const showSearchEmptyState =
     isTableEmpty && !!searchCel && !showFilterEmptyState;
@@ -552,11 +546,9 @@ export function AlertTableServerSide({
                     <AlertsTableBody
                       table={table}
                       showSkeleton={showSkeleton}
-                      showEmptyState={showEmptyState}
                       showFilterEmptyState={showFilterEmptyState}
                       showSearchEmptyState={showSearchEmptyState}
                       theme={theme}
-                      viewedAlerts={viewedAlerts}
                       lastViewedAlert={lastViewedAlert}
                       onRowClick={handleRowClick}
                       onClearFiltersClick={() => setClearFiltersToken(uuidV4())}
